requirements.txt
fastapi
uvicorn
joblib
numpy
torch
pandas
scikit-learn
pydantic
langchain
llama-cpp-python
langchain-community
replicate
sqlalchemy
psycopg2